# api/API/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api de blog du cours Api",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "sandbox-server"
        }
    ],
    "paths": {
...
        "/produit": {
            "get": {
                "summary": "List all products",
                "description": "List all products",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit permettant la pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List all products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Produit[]",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Produit"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "apiKey missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/produit/{id}": {
            "get": {
                "summary": "return a produiuct",
                "description": "return a product",
                "operationId": "abcbc550c1a09c8c10800255187236f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du produit à récupérer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return a product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Produit"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "apiKey missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/produit/create": {
            "post": {
                "summary": "Create a product",
                "description": "Save product in DB",
                "operationId": "ddf41842d908a155aefc396a46453626",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Produit à enregistrer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Collier rouge"
                                    },
                                    "cible": {
                                        "type": "string",
                                        "example": "Chiens"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Accessoires"
                                    },
                                    "prix": {
                                        "type": "number",
                                        "example": 5.5
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Collier rouge largeur 2 cm"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "https://www.wanimo.com/images/collier_laisse_et_harnais/A/SM/ASMS040/500x500/01/collier-basic-en-cuir-rouge-martin-sellier.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produit enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/produit/{id}/update": {
            "put": {
                "summary": "Update product",
                "description": "Update product in Db",
                "operationId": "5c1e704b31058f22d39dab48c39cd361",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du produit à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Nouvelles données du produit",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Collier rouge"
                                    },
                                    "cible": {
                                        "type": "string",
                                        "example": "Chiens"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Accessoire"
                                    },
                                    "prix": {
                                        "type": "double",
                                        "example": 5.5
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Collier rouge largeur 2 cm"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "https://www.wanimo.com/images/collier_laisse_et_harnais/A/SM/ASMS040/500x500/01/collier-basic-en-cuir-rouge-martin-sellier.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produit enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/produit/{id}/delete": {
            "delete": {
                "summary": "Delete product",
                "description": "Delete product in Db",
                "operationId": "186a1916d4649a776e3188d4e57e4cf0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du produit à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suppression validée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Produit supprimé"
                            }
                        }
                    }
                }
            }
        },
...
    }
}